This request proposes to change `FileChannel.transferFrom()` to add a fast path on Linux and a threshold for mapped transfers on all platforms.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed (1 review required, with at least 1 reviewer)

### Issue
 * [JDK-8274113](https://bugs.openjdk.java.net/browse/JDK-8274113): (fc) Tune FileChannel.transferFrom()


### Reviewers
 * [Alan Bateman](https://openjdk.java.net/census#alanb) (@AlanBateman - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/8644/head:pull/8644` \
`$ git checkout pull/8644`

Update a local copy of the PR: \
`$ git checkout pull/8644` \
`$ git pull https://git.openjdk.java.net/jdk pull/8644/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 8644`

View PR using the GUI difftool: \
`$ git pr show -t 8644`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/8644.diff">https://git.openjdk.java.net/jdk/pull/8644.diff</a>

</details>